Moved the service calls to the plugin
Moved the calls to Umbraco's `editorService` API away from the main
script over to be the responsibility of the UmbracoMedia (ImageTool)
plugin itself.

This puts less dependency on the property-editor's script.

In doing this, we could also relocate the ImageTool's resources API
call too.

I'm not completely happy with the `requiresEditorService: true` config.
I'd have preferred a way to use DI to inject any Umbraco/AngularJs
service, but my AngularJs knowledge doesn't stretch that far! :wink: